fix: move to sub folder and change controlledValue prop
