+++ packages/design-system/src/components/Inputs/index.ts
@@ -1,1 +1,1 @@
-export { default as Input } from './Input';
+export * from './Input';
